requirements.txt
asgiref==3.8.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
bootstrap4==0.1.0
decorators==2.0.7
dj-database-url==2.2.0
Django==5.0.6
django-bootstrap4==24.3
djangorestframework==3.15.1
gunicorn==22.0.0
packaging==24.0
psycopg==3.1.19
psycopg2==2.9.9
psycopg2-binary==2.9.9
soupsieve==2.5
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
utils==1.0.2
whitenoise==6.6.0
